# services/flight/api/openapi.yaml
openapi: 3.0.3
info:
  title: FlightService API
  description: Service zur Verwaltung von Flugbuchungen
  version: 1.0.0
servers:
  - url: http://localhost/api/flight
    description: API-Gateway (Traefik)
paths:
...
  /bookings:
    post:
      summary: Flug buchen
      description: |
        Bestätigt eine Flugbuchung und loggt sie auf stdout. Es findet keine
        Persistenz statt — die Buchung existiert nur in den Logs.
      operationId: createBooking
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BookingRequest'
      responses:
        '201':
          description: Buchung erfolgreich bestätigt
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Booking'
        '400':
          description: Ungültige Anfrage
...
  /events/compensation:
    post:
      summary: Kompensations-Event entgegennehmen (Story 7)
      description: |
        Nimmt ein `CompensationRequested`-Event vom Booking-Service entgegen
        und antwortet **sofort mit 202 Accepted**. Die eigentliche Stornierung
        läuft asynchron in einer Goroutine — der Sender wartet nicht auf das
        Ergebnis (fire & forget). Genutzt in der Choreography-Saga von Story 7,
        im Gegensatz zum synchronen `DELETE /bookings/{id}` aus Story 6.
      operationId: receiveCompensationEvent
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CompensationEvent'
      responses:
        '202':
          description: Event entgegengenommen, Verarbeitung läuft asynchron
        '400':
          description: Ungültiges Event (fehlende Pflichtfelder)
components:
  schemas:
...
    BookingRequest:
      type: object
      required:
        - flightId
        - customerName
      properties:
        flightId:
          type: string
          example: LH123
        customerName:
          type: string
          example: Max Mustermann
    Booking:
      type: object
      properties:
        bookingId:
          type: string
          example: F-7c1a9f
        flightId:
          type: string
          example: LH123
        customerName:
          type: string
          example: Max Mustermann
        status:
          type: string
          enum: [CONFIRMED]
          example: CONFIRMED
    CompensationEvent:
      type: object
      required:
        - eventId
        - sagaId
        - bookingId
      properties:
        eventId:
          type: string
          description: Eindeutige Event-ID, vom Sender erzeugt. Nutzbar als Dedup-Key für Idempotenz.
          example: ev-3f9a1c
        sagaId:
          type: string
          description: Saga-ID des Booking-Service, mit der das Event korreliert wird.
          example: S-7c1a9f
        bookingId:
          type: string
          description: ID der zu stornierenden Buchung in diesem Service.
          example: F-7c1a9f